+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All changes will be documented here, with releases being tagged + linked.
 
+## Unreleased
+
+- Support `scale` transform option
+
 ## 0.1.1
 
 - #2 - Form editor now correctly pushes other row entries down the grid